add email send

+++ src/task/controller.py
@@ -1,6 +1,6 @@
 from fastapi import Depends,HTTPException
 from sqlalchemy.orm import Session
-from src.task.dtos import TaskSchema
+from src.task.dtos import TaskSchema, UpdateTaskSchema
 from src.task.model import Task
 from src.utils.db import get_db
 from src.user.model import User
@@ -30,10 +30,13 @@ def get_one_task(id: int, db: Session, user_id: int):
         raise HTTPException(status_code=404, detail="Task not found")
     return task
 
-def update_task(id: int, body: TaskSchema, db: Session, user_id: int):
-    task = db.query(Task).filter(Task.id == id, Task.user_id == user_id).first()
+def update_task(id: int, body: UpdateTaskSchema, db: Session, user_id: int):
+    task = db.query(Task).filter(Task.id == id).first()
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
+    if task.user_id != user_id:
+        raise HTTPException(status_code=401, detail="You are not allowed to update this")
+    
     task.title = body.title
     task.description = body.description
     task.status = body.status
@@ -42,9 +45,11 @@ def update_task(id: int, body: TaskSchema, db: Session, user_id: int):
     return task
 
 def delete_task(id: int, db: Session, user_id: int):
-    task = db.query(Task).filter(Task.id == id, Task.user_id == user_id).first()
+    task = db.query(Task).filter(Task.id == id).first()
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
+    if task.user_id != user_id:
+        raise HTTPException(status_code=401, detail="You are not allowed to delete this")
     db.delete(task)
     db.commit()
     return task
